# swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"Nieuwkoop Europe B.V. Customer ERP API V1.1","description":"For documentation on usage please go to our website <a href='https://developer.nieuwkoop-europe.io/' target='_blank'>Documentation (EN)</a>"},"host":"customerapi.nieuwkoop-europe.com","schemes":["https"],"paths":{"/items":{"get":{"tags":["Items"],"summary":"Get the items modified since {sysmodified}. Include {itemCode} to only get the specified item.","operationId":"Items_GetItems","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"sysmodified","in":"query","description":"Date and time of last modification.","required":true,"type":"string","format":"date-time"},{"name":"itemCode","in":"query","description":"Itemcode","required":false,"type":"string"},{"name":"mainGroupCode","in":"query","description":"MainGroupCode","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/ItemViewModel"}}}}}},"/items/{itemCode}/image":{"get":{"tags":["Items"],"operationId":"Items_GetImage","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"itemCode","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ItemImage"}}}}},"/stock":{"get":{"tags":["Items"],"operationId":"Items_GetStock","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"sysmodified","in":"query","required":true,"type":"string","format":"date-time"},{"name":"itemCode","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Stock"}}}}}},"/tags":{"get":{"tags":["Items"],"summary":"Get a list of tags and their possible values. Include {tagCode} to only get the specified tag.","operationId":"Items_GetTag","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"tagCode","in":"query","description":"(Optional) TagCode to get.","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/TagViewModel"}}}}}},"/prices":{"get":{"tags":["Items"],"summary":"Get item prices modified since {sysmodified}. Include {itemCode} to only get the price for a single item.","operationId":"Items_GetItemPrices","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"sysmodified","in":"query","description":"Date and time of last modification.","required":true,"type":"string","format":"date-time"},{"name":"itemCode","in":"query","description":"(Optional) ItemCode","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/ItemPriceViewModel"}}}}}},"/salesorders":{"get":{"tags":["SalesOrders"],"summary":"Get the orders modified since {sysmodified}.\r\n\r\nInclude values for {orderNr}, {externalReference}, {reference}, {yourReference},\r\n{accountCode}, {order_ContactName} and/or {delivered} to perform additional filtering.\r\n\r\nSpecify {includeLines} to either include (default) or exclude SalesOrderLines in the response.","description":"This endpoint supports sorting (with the {sorting} parameter) and returns a paged result (maximum of {pageSize} items per request).\r\n\r\nThe response will contain HTTP Headers for the paging information:\r\n* `X-Total-Count` The total number of orders that match the filters.\r\n* `X-Current-Page` The page number of the requested page.\r\n* `X-Results-Per-Page` The page size.\r\n\r\nThe {sorting} is a comma-separated list of the field(s) to sort the result on (with optional direction per field).\r\n\r\nYou can specify the sort direction by either prepending the field with `+` or `-` (ascending or descending) or by adding `ASC` or `DESC`. For instance:\r\n* `Sysmodified,OrderNr`: Default sorting, first by Sysmodified then by OrderNr\r\n* `-DeliveryDate,OrderNr` or `DeliveryDate DESC,OrderNr`: DeliveryDate descending, then by orderNr\r\n* `-OrderNr` or `OrderNr DESC`: OrderNr descending\r\n            \r\nTo guarantee a deterministic order of the items, the OrderNr field will always be included as the final sorted field.","operationId":"SalesOrders_GetSalesOrders","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"sysmodified","in":"query","description":"(Optional) Date and time of last modification.","required":false,"type":"string","format":"date-time"},{"name":"orderNr","in":"query","description":"(Optional) Order number","required":false,"type":"string"},{"name":"externalReference","in":"query","description":"(Optional) External reference","required":false,"type":"string"},{"name":"reference","in":"query","description":"(Optional) Reference","required":false,"type":"string"},{"name":"yourReference","in":"query","description":"(Optional) YourReference","required":false,"type":"string"},{"name":"accountCode","in":"query","description":"(Optional) Account code","required":false,"type":"string"},{"name":"order_ContactName","in":"query","description":"(Optional) Contact name from the order address.","required":false,"type":"string"},{"name":"delivered","in":"query","description":"(Optional) delivered; when set only return (not) delivered orders.","required":false,"type":"boolean"},{"name":"includeLines","in":"query","description":"(Optional, default \"true\") Include the sales order lines in the response.","required":false,"type":"boolean"},{"name":"pageSize","in":"query","description":"(Optional, default 50) the page size","required":false,"type":"integer","format":"int32"},{"name":"pageNumber","in":"query","description":"(Optional, default 1) the page number","required":false,"type":"integer","format":"int32"},{"name":"sorting","in":"query","description":"(Optional, default Sysmodified,OrderNr) sorting (see Implementation notes)","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/SalesOrderRead"}}}}}},"/salesorders/contacts/{accountCode}":{"get":{"tags":["SalesOrders"],"summary":"Get the different \"Order_ContactName\" values that are used in orders for a specific {accountCode}.\r\n\r\nYou can optionally filter with {delivered} to only include orders that are (not) delivered.","operationId":"SalesOrders_GetOrderContacts","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"accountCode","in":"path","description":"The accountCode used in orders to get the contacts for.","required":true,"type":"string"},{"name":"delivered","in":"query","description":"(Optional) filter to only get contacts used in (not) delivered orders.","required":false,"type":"boolean"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"string"}}}}}},"/salesorders/yourreferences/{accountCode}":{"get":{"tags":["SalesOrders"],"summary":"Get the different \"Yourreference\" values that are used in orderlines for orders for a specific {accountCode}.\r\n\r\nYou can optionally filter with {delivered} to only include orders that are (not) delivered.","operationId":"SalesOrders_GetOrderYourreferences","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"accountCode","in":"path","description":"The accountCode used in orders to get the Yourreferences for.","required":true,"type":"string"},{"name":"delivered","in":"query","description":"(Optional) filter to only get yourreferences from (not) delivered orders.","required":false,"type":"boolean"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"string"}}}}}},"/salesorder":{"post":{"tags":["SalesOrders"],"summary":"Create a new sales order.","operationId":"SalesOrders_SalesOrder","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"salesOrder","in":"body","description":"Sales order header and lines","required":true,"schema":{"$ref":"#/definitions/SalesOrder"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CreateSalesOrderResponse"}}}}},"/users/passwordResetRequest":{"post":{"tags":["Users"],"operationId":"Users_PasswordResetRequest","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":[],"parameters":[{"name":"passwordResetRequest","in":"body","required":true,"schema":{"$ref":"#/definitions/PasswordResetRequest"}}],"responses":{"204":{"description":"No Content"}}}}},"definitions":{"ItemViewModel":{"type":"object","properties":{"Itemcode":{"type":"string"},"DeliveryTimeInDays":{"format":"int32","type":"integer"},"PalletQuantity":{"format":"double","type":"number"},"Description":{"type":"string"},"ItemDescription_NL":{"type":"string"},"ItemStatus":{"type":"string"},"SalesPackage_NL":{"type":"string"},"SalesPackage_DE":{"type":"string"},"SalesPackage_EN":{"type":"string"},"SalesPackage_FR":{"type":"string"},"Salesprice":{"format":"double","type":"number"},"MainGroupCode":{"type":"string"},"MainGroupDescription_NL":{"type":"string"},"MainGroupDescription_DE":{"type":"string"},"MainGroupDescription_EN":{"type":"string"},"MainGroupDescription_FR":{"type":"string"},"ProductGroupCode":{"type":"string"},"ProductGroupDescription_NL":{"type":"string"},"ProductGroupDescription_DE":{"type":"string"},"ProductGroupDescription_EN":{"type":"string"},"ProductGroupDescription_FR":{"type":"string"},"GroupDescription":{"type":"string"},"GroupDescription_NL":{"type":"string"},"GroupDescription_DE":{"type":"string"},"GroupDescription_EN":{"type":"string"},"GroupDescription_FR":{"type":"string"},"MaterialGroupCode":{"type":"string"},"MaterialGroupDescription_NL":{"type":"string"},"MaterialGroupDescription_DE":{"type":"string"},"MaterialGroupDescription_EN":{"type":"string"},"MaterialGroupDescription_FR":{"type":"string"},"ItemDescription_EN":{"type":"string"},"ItemDescription_DE":{"type":"string"},"ItemDescription_FR":{"type":"string"},"GTINCode":{"type":"string"},"IsStockItem":{"type":"boolean"},"Warehouse":{"type":"string"},"ItemVariety_NL":{"type":"string"},"ItemVariety_DE":{"type":"string"},"ItemVariety_EN":{"type":"string"},"ItemVariety_FR":{"type":"string"},"PotSize":{"type":"string"},"ItemPictureName":{"type":"string"},"ItemPictureSysmodified":{"format":"date-time","type":"string"},"Content_Ltr":{"type":"string"},"PlantPassportCode":{"type":"string"},"Diameter":{"format":"double","type":"number"},"Length":{"format":"double","type":"number"},"Width":{"format":"double","type":"number"},"Height":{"format":"double","type":"number"},"Depth":{"format":"double","type":"number"},"Opening":{"format":"double","type":"number"},"IsOffer":{"type":"boolean"},"ShowOnWebsite":{"type":"boolean"},"Sysmodified":{"format":"date-time","type":"string"},"SalesOrderSize":{"format":"int32","type":"integer"},"DiameterCulturePot":{"format":"double","type":"number"},"HeightCulturePot":{"format":"double","type":"number"},"Weight":{"format":"double","type":"number"},"CitesListed":{"type":"boolean"},"CountryOfOrigin":{"type":"string"},"CountryOfProvenance":{"type":"string"},"FytoListed":{"type":"boolean"},"HSCode":{"format":"int32","type":"integer"},"HSCodeUK":{"type":"string"},"LeafSize":{"type":"string"},"QuantityPallet":{"format":"int32","type":"integer"},"QuantityTrolley":{"format":"int32","type":"integer"},"LocationIcon_DE":{"type":"string"},"LocationIcon_EN":{"type":"string"},"LocationIcon_FR":{"type":"string"},"LocationIcon_NL":{"type":"string"},"LocationUsagePlanters_DE":{"type":"string"},"LocationUsagePlanters_EN":{"type":"string"},"LocationUsagePlanters_FR":{"type":"string"},"LocationUsagePlanters_NL":{"type":"string"},"onlyShipTo":{"type":"string"},"PEC":{"format":"int32","type":"integer"},"isProduction":{"format":"int32","type":"integer"},"Tags":{"type":"array","items":{"$ref":"#/definitions/ItemTagViewModel"}}}},"ItemTagViewModel":{"type":"object","properties":{"Code":{"type":"string"},"Values":{"type":"array","items":{"$ref":"#/definitions/TagValueViewModel"}}}},"TagValueViewModel":{"type":"object","properties":{"Description_NL":{"type":"string"},"Description_DE":{"type":"string"},"Description_EN":{"type":"string"},"Description_FR":{"type":"string"}}},"ItemImage":{"type":"object","properties":{"Itemcode":{"type":"string"},"Image":{"format":"byte","type":"string"},"Sysmodified":{"format":"date-time","type":"string"}}},"Stock":{"type":"object","properties":{"Itemcode":{"type":"string"},"StockAvailable":{"format":"double","type":"number"},"FirstAvailable":{"format":"date-time","type":"string"},"Sysmodified":{"format":"date-time","type":"string"}}},"TagViewModel":{"type":"object","properties":{"TagCode":{"type":"string"},"Description_NL":{"type":"string"},"Description_DE":{"type":"string"},"Description_EN":{"type":"string"},"Description_FR":{"type":"string"},"WebDescription_NL":{"type":"string"},"WebDescription_DE":{"type":"string"},"WebDescription_EN":{"type":"string"},"WebDescription_FR":{"type":"string"},"Values":{"type":"array","items":{"$ref":"#/definitions/TagValueViewModel"}}}},"ItemPriceViewModel":{"type":"object","properties":{"ItemCode":{"type":"string"},"PriceGross":{"format":"double","type":"number"},"DiscountPercentage":{"format":"double","type":"number"},"PriceNett":{"format":"double","type":"number"},"Staff1":{"$ref":"#/definitions/StaffelViewModel"},"Staff2":{"$ref":"#/definitions/StaffelViewModel"},"Staff3":{"$ref":"#/definitions/StaffelViewModel"},"Staff4":{"$ref":"#/definitions/StaffelViewModel"},"Staff5":{"$ref":"#/definitions/StaffelViewModel"},"Sysmodified":{"format":"date-time","type":"string"}}},"StaffelViewModel":{"type":"object","properties":{"Quantity":{"format":"double","type":"number"},"Price":{"format":"double","type":"number"}}},"SalesOrderRead":{"type":"object","properties":{"Division":{"type":"string"},"OrderNr":{"type":"string"},"ExternalReference":{"type":"string"},"Reference":{"type":"string"},"AccountCode":{"type":"string"},"GLN":{"type":"string"},"OrderDate":{"format":"date-time","type":"string"},"DeliveryDate":{"format":"date-time","type":"string"},"DeliveryMethod":{"type":"string"},"PaymentCondition":{"type":"string"},"Notes":{"type":"string"},"Shipping_AccountCode":{"type":"string"},"Shipping_GLN":{"type":"string"},"Shipping_CompanyName":{"type":"string"},"Shipping_ContactName":{"type":"string"},"Shipping_Address1":{"type":"string"},"Shipping_Address2":{"type":"string"},"Shipping_PostalCode":{"type":"string"},"Shipping_City":{"type":"string"},"Shipping_Country":{"type":"string"},"Invoice_AccountCode":{"type":"string"},"Invoice_GLN":{"type":"string"},"Invoice_CompanyName":{"type":"string"},"Invoice_ContactName":{"type":"string"},"Invoice_Address1":{"type":"string"},"Invoice_Address2":{"type":"string"},"Invoice_PostalCode":{"type":"string"},"Invoice_City":{"type":"string"},"Invoice_Country":{"type":"string"},"Syscreated":{"format":"date-time","type":"string"},"Sysmodified":{"format":"date-time","type":"string"},"Delivered":{"type":"boolean"},"Order_AccountCode":{"type":"string"},"Order_CompanyName":{"type":"string"},"Order_ContactName":{"type":"string"},"Order_Address1":{"type":"string"},"Order_Address2":{"type":"string"},"Order_PostalCode":{"type":"string"},"Order_City":{"type":"string"},"Order_Country":{"type":"string"},"OrderCompanyRepresentative":{"type":"string"},"Currency":{"type":"string"},"Source":{"type":"string"},"OrderApproveStatus":{"type":"string"},"DeadlineApproval":{"format":"date-time","type":"string"},"DeadlineDelivery":{"format":"date-time","type":"string"},"CarrierCC":{"format":"int32","type":"integer"},"CarrierFH":{"format":"int32","type":"integer"},"CarrierBlock":{"format":"int32","type":"integer"},"CarrierEuro":{"format":"int32","type":"integer"},"CarrierDC":{"format":"int32","type":"integer"},"CarrierSeperate":{"type":"boolean"},"OrderStatus":{"type":"string"},"AssemblyOrder":{"type":"boolean"},"DeadlineDeliveryYN":{"type":"boolean"},"InvoiceNr":{"type":"string"},"InvoiceStatus":{"type":"string"},"TotalGross":{"format":"double","type":"number"},"TotalDiscount":{"format":"double","type":"number"},"TotalVAT":{"format":"double","type":"number"},"TotalNetVAT":{"format":"double","type":"number"},"SalesOrderLines":{"type":"array","items":{"$ref":"#/definitions/SalesOrderLineRead"}}}},"SalesOrderLineRead":{"type":"object","properties":{"OrderNr":{"type":"string"},"LineID":{"format":"int32","type":"integer"},"Itemcode":{"type":"string"},"GTINCode":{"type":"string"},"Description":{"type":"string"},"Quantity":{"format":"double","type":"number"},"Price":{"format":"double","type":"number"},"Discount":{"format":"double","type":"number"},"NetPrice":{"format":"double","type":"number"},"DeliveryDate":{"format":"date-time","type":"string"},"Instruction":{"type":"string"},"Extra1":{"type":"string"},"Extra2":{"type":"string"},"Extra3":{"type":"string"},"Syscreated":{"format":"date-time","type":"string"},"Sysmodified":{"format":"date-time","type":"string"},"Origine":{"type":"string"},"Yourreference":{"type":"string"},"DeliverStatus":{"type":"boolean"},"SprayColour":{"type":"string"},"AssemblyGroup":{"type":"string"},"VATCode":{"type":"string"},"OrderlineApproveStatus":{"type":"string"},"Variety_NL":{"type":"string"},"Variety_DE":{"type":"string"},"Variety_FR":{"type":"string"},"Variety_EN":{"type":"string"},"Diameter":{"format":"double","type":"number"},"Length":{"format":"double","type":"number"},"Width":{"format":"double","type":"number"},"Height":{"format":"double","type":"number"},"Description_NL":{"type":"string"},"Description_DE":{"type":"string"},"Description_FR":{"type":"string"},"Description_EN":{"type":"string"}}},"SalesOrder":{"type":"object","properties":{"Division":{"type":"string"},"ExternalReference":{"type":"string"},"Reference":{"type":"string"},"AccountCode":{"type":"string"},"GLN":{"type":"string"},"OrderDate":{"format":"date-time","type":"string"},"DeliveryDate":{"format":"date-time","type":"string"},"DeliveryMethod":{"type":"string"},"PaymentCondition":{"type":"string"},"Notes":{"type":"string"},"Shipping_AccountCode":{"type":"string"},"Shipping_GLN":{"type":"string"},"Shipping_CompanyName":{"type":"string"},"Shipping_ContactName":{"type":"string"},"Shipping_Address1":{"type":"string"},"Shipping_Address2":{"type":"string"},"Shipping_PostalCode":{"type":"string"},"Shipping_City":{"type":"string"},"Shipping_Country":{"type":"string"},"Invoice_AccountCode":{"type":"string"},"Invoice_GLN":{"type":"string"},"Invoice_CompanyName":{"type":"string"},"Invoice_ContactName":{"type":"string"},"Invoice_Address1":{"type":"string"},"Invoice_Address2":{"type":"string"},"Invoice_PostalCode":{"type":"string"},"Invoice_City":{"type":"string"},"Invoice_Country":{"type":"string"},"SalesOrderLines":{"type":"array","items":{"$ref":"#/definitions/SalesOrderLine"}}}},"SalesOrderLine":{"type":"object","properties":{"LineID":{"format":"int32","type":"integer"},"Itemcode":{"type":"string"},"GTINCode":{"type":"string"},"Description":{"type":"string"},"Quantity":{"format":"double","type":"number"},"Price":{"format":"double","type":"number"},"Discount":{"format":"double","type":"number"},"NetPrice":{"format":"double","type":"number"},"DeliveryDate":{"format":"date-time","type":"string"},"Instruction":{"type":"string"},"Extra1":{"type":"string"},"Extra2":{"type":"string"},"Extra3":{"type":"string"}}},"CreateSalesOrderResponse":{"type":"object","properties":{"OrderNr":{"type":"string"},"Status":{"type":"string"},"ErrorMessage":{"type":"string"}}},"PasswordResetRequest":{"required":["EmailAddress","ResetLink"],"type":"object","properties":{"EmailAddress":{"type":"string"},"ResetLink":{"type":"string"},"Type":{"type":"string"},"Key":{"type":"string"},"Data":{"$ref":"#/definitions/PasswordResetRequestData"}}},"PasswordResetRequestData":{"type":"object","properties":{"Pdf":{"type":"boolean"},"Xls":{"type":"boolean"},"Datasheet":{"type":"boolean"},"Images":{"type":"boolean"},"Proforma":{"format":"int32","type":"integer"}}}},"securityDefinitions":{"basic":{"type":"basic","description":"Basic HTTP Authentication"}}}